+++ app/src/main/java/nextstep/payments/ui/screen/newcard/NewCardScreen.kt
@@ -20,7 +20,7 @@ import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
 import nextstep.payments.ui.component.NewCardTopBar
-import nextstep.payments.ui.component.PaymentCard
+import nextstep.payments.ui.component.card.PaymentCard
 import nextstep.payments.ui.component.text.CreditCardVisualTransformation
 import nextstep.payments.ui.component.text.ExpirationDateVisualTransformation
 import nextstep.payments.ui.theme.PaymentsTheme
